renamed debug folder to utils and added some extra utilities for the game to aid development
