--- debug/data/wordsets/science.txt
@@ -1,587 +1,0 @@
-Skeleton
-Bone
-Bones
-Skin
-Brain
-Brains
-Finger
-Fingers
-Hand
-Hands
-Arm
-Arms
-Knee
-Knees
-Foot
-Feet
-Leg
-Legs
-Toe
-Toes
-Thumb
-Thumbs
-Limb
-Limbs
-Spine
-Health
-Plants
-Force
-Growth
-Material
-Magnet
-Springs
-Living
-Habitats
-Variation
-Electricity
-Sound
-Light
-Earth
-Space
-Neptune
-Saturn
-Jupiter
-Mercury
-Venus
-Uranus
-Sun
-Pluto
-Mars
-Rock
-Soil
-Adaptation
-Lifecycle
-Insulator
-Conductor
-Thermal
-Fish
-Reptiles
-Mammals
-Birds
-Amphibians
-Herbivore
-Omnivore
-Carnivore
-Leg
-Arm
-Elbow
-Head
-Ear
-Nose
-Back
-Leaves
-Bulb
-Branches
-Stem
-Nutrition
-Air
-Water
-Nutrients
-Reproduction
-Pollinate
-Pollination
-Flower
-Crystals
-Marble
-Granite
-Dead
-Energy
-Chain
-Predator
-Prey
-Desert
-Pond
-River
-Lake
-Sea
-Ocean
-Temperature
-Survive
-Survival
-Rough
-Smooth
-Metal
-Solid
-Gas
-Liquid
-Condensation
-Particle
-Particles
-Freeze
-Freezing
-Hot
-Cold
-Heat
-Boil
-Mouth
-Tongue
-Teeth
-Stomach
-Intestine
-Transparent
-Filter
-Blood
-Oxygen
-Carbon
-Respiration
-Exercise
-Valve
-Veins
-Artery
-Arteries
-Temperature
-Element
-Elements
-Periodic
-Biology
-Chemistry
-Physics
-Chemical
-Chemicals
-Bacteria
-Bacterium
-Disease
-Virus
-Microbe
-Digest
-Digestion
-Oxidation
-Rust
-Fossil
-Fuel
-Photosynthesis
-Ion
-Photon
-Photons
-Electron
-Electron
-Neutron
-Neutrons
-Atom
-Atoms
-Star
-Stars
-Nebula
-Nebulae
-Galaxy
-Galaxies
-Gravity
-Moon
-Moons
-Constellation
-Constellation
-Astrology
-Astronomy
-Astronaut
-Astronauts
-Rocket
-Satelite
-Geology
-Geologist
-Geo
-Geode
-Electrode
-Combustion
-Ignition
-Flammable
-Fire
-Carbonize
-Carbonized
-Liquefaction
-Emulsion
-Iron
-Zinc
-Gold
-Silver
-Bronze
-Platinum
-Titanium
-Helium
-Hydrogen
-Lithium
-Nitrogen
-Argon
-Cobalt
-Chromium
-Potassium
-Barium
-Copper
-Iridium
-Mercury
-Nickel
-Tungsten
-Sodium
-Alpha
-Beta
-Gamma
-Acceleration
-Accelerate
-Accelerating
-Accelerates
-Accelerated
-Displacement
-Displace
-Displaced
-Displaces
-Displacing
-Vector
-Speed
-Distance
-Time
-Formula
-Formulas
-Formulae
-Radiation
-Toxic
-Poison
-Corrosion
-Corrosive
-Acid
-Acids
-Acidic
-Alkali
-Limestone
-Composition
-Decomposition
-Decompose
-Decomposes
-Decomposed
-Decomposing
-Decay
-Decays
-Decayed
-Decays
-Decaying
-Constant
-Variable
-Variables
-Watt
-Watts
-Second
-Seconds
-Minute
-Minutes
-Hour
-Hours
-Day
-Days
-Week
-Weeks
-Month
-Months
-Year
-Years
-Decade
-Decades
-Century
-Centuries
-Millenium
-Milleniums
-Electrocute
-Spark
-Friction
-Weight
-Density
-Theory
-Test
-Thesis
-Analysis
-Matter
-Carbon
-Carbohydrate
-Hydrate
-Hydration
-Dehydration
-Dinosaur
-Magnetism
-Calcium
-Protein
-Fat
-Vitamin
-Vitamins
-Mineral
-Minerals
-Halogen
-Metabolism
-Nuclear
-Solar
-Wind
-Turbine
-Engine
-Accelerator
-Evolve
-Evolution
-Darwin
-Natural
-Selection
-Seed
-Seeds
-Muscle
-Muscles
-Joint
-Joints
-Synapse
-Wave
-Waves
-Wavelength
-Frequency
-Frequencies
-Amplitude
-Crest
-Crests
-Peak
-Peaks
-Trough
-Troughs
-Transverse
-Pressure
-Period
-Oscillation
-Oscillate
-Opacity
-Refract
-Refraction
-Reflect
-Reflection
-Property
-Micro
-Nano
-Mile
-Miles
-Kilometer
-Kilometers
-Meter
-Meters
-Centimeter
-Centimeters
-Millimeter
-Millimeters
-Kilogram
-Kilograms
-Gram
-Grams
-Milligram
-Milligrams
-Degrees
-Fahrenheit
-Kelvin
-Mass
-Current
-Luminous
-Luminosity
-Length
-Cell
-Cells
-Membrane
-Nucleus
-Cytoplasm
-Ribosome
-Mitochondria
-Vacuole
-Core
-Dissect
-Dissection
-Drug
-Drugs
-Medicine
-Medicines
-Vaccine
-Antibody
-Antibodies
-Penicillin
-Infection
-Infections
-Infected
-Disinfect
-Disinfected
-Disinfection
-Mold
-Vapour
-Vapourise
-Vaporizing
-Vaporized
-Evaporate
-Evaporation
-Evaporating
-Evaporated
-Sublimation
-Sublimating
-Sublimated
-Deposition
-Disposed
-Melt
-Melting
-Melted
-Rigid
-State
-States
-Equation
-Ozone
-Atmosphere
-Meteor
-Comet
-Asteroid
-Crust
-Mantle
-Inorganic
-Organic
-Reactive
-Explosion
-Explosive
-Nobel
-Boiling
-Boiled
-Science
-Scientist
-Doctor
-Nurse
-Pharmacy
-Pharmacist
-Biologist
-Chemist
-Physicist
-Wildlife
-Marine
-Lung
-Lungs
-Heart
-Hearts
-Bladder
-Bladders
-Liver
-Livers
-Kidney
-Kidneys
-Vibrate
-Vibration
-Structure
-Transform
-Transformation
-Transformed
-Transforming
-Transforms
-Mutate
-Mutation
-Mutating
-Mutates
-Mutated
-Artificial
-Lifeform
-Genetic
-Genetics
-Clone
-Clones
-Cloning
-Chromosome
-Chromosomes
-Male
-Female
-Reproduce
-Reproduced
-Reproduction
-Sex
-Sexual
-Sexually
-Assexually
-Egg
-Fertilise
-Fertilises
-Fertilising
-Fertilisation
-Fertilised
-Fetus
-Embryo
-Zygote
-Abortion
-Amp
-Amps
-Multicellular
-Body
-Eye
-Eyes
-Ear
-Ears
-Hair
-Hairs
-Heel
-Filament
-Anatomy
-Gastric
-Sweat
-Sweats
-Perspiration
-Perspirate
-Perspirates
-Aspirin
-Painkiller
-Painkillers
-Stress
-Cancer
-Diabetes
-Cure
-Alcohol
-Spirit
-Spirits
-Distill
-Distills
-Distilling
-Distillation
-Distilated
-Filter
-Filters
-Filtering
-Filtration
-Filtered
-Activate
-Activates
-Activated
-Ferment
-Ferments
-Fermenting
-Fermentation
-Fermented
-Battery
-Batteries
-Power
-Powered
-Powering
-Node
-Bulb
-Wire
-Wires
-Wired
-Disrupt
-Disruption
-Disrupted
-Disrupting
-Conducting
-Conducts
-Conduct
-Insulate
-Insulates
-Insulating
-Insulated
-Molecule
-Molecular
-Condense
-Condensed
-Disperse
-Dispersion
-Dispersed
-Dispersing
-Diffuse
-Disfuses
-Diffusion
-Defused
-Disfusing
-Resistance
-Resist
-Resists
-Resisting
-Resisted
-Food
-Producer
-Producers
-Consumer
-Consumers
-Nutrition
