+++ utils/data/wordsets/maths.txt
@@ -1,0 +1,322 @@
+Add
+Addition
+Minus
+Subtract
+Subtraction
+Divide
+Multiply
+Multiplication
+Multiplicative
+Times
+Product
+Number
+Numeral
+Figure
+Value
+Zero
+One
+Two
+Three
+Four
+Five
+Six
+Seven
+Eight
+Nine
+Ten
+Eleven
+Twelve
+Thirteen
+Fourteen
+Fifteen
+Sixteen
+Seventeen
+Eighteen
+Nineteen
+Twenty
+Thirty
+Fourty
+Fifty
+Sixty
+Seventy
+Eighty
+Ninety
+Hundred
+Hundred
+Thousand
+Thousands
+Million
+Millions
+Billion
+Billions
+Trillion
+Trillions
+Decimal
+Decimals
+Fraction
+Fractions
+Percentage
+Percentages
+Multiples
+Factorial
+Power
+Powers
+Root
+Index
+Indices
+Surd
+Surds
+Ratio
+Angle
+Proportion
+Statistic
+Statistics
+Statistical
+Anomaly
+Probable
+Probability
+Coefficient
+Algebra
+Square
+Prime
+Proportion
+Multiples
+Shape
+Triangle
+Triangular
+Angular
+Pentagon
+Pentagonal
+Hexagon
+Hexagonal
+Octagon
+Octagonal
+Trapezium
+Vertex
+Symmetrical
+Rotate
+Rotation
+Parallelogram
+Oblique
+Acute
+Average
+Dodecagon
+Adjacent
+Intersection
+Even
+Odd
+Alternate
+Area
+Axis
+Digit
+Edge
+Edges
+Equation
+Face
+Faces
+Factor
+Factors
+Greater
+Less
+Integer
+Integers
+Float
+Floats
+Floating
+Mean
+Mode
+Median
+Heptagon
+Heptagonal
+Isosceles
+Kite
+Rhombus
+Horizontal
+Vertical
+Numerator
+Oblong
+Perimeter
+Parallel
+Line
+Lines
+Perpendicular
+Quadrant
+Rectangle
+Rectangular
+Reflex
+Round
+Rounding
+Scalene
+Score
+Sum
+Tally
+Vertex
+Vertices
+Exponent
+Exponential
+Linear
+Quadratic
+Cubic
+Difference
+Differentiation
+Integration
+Derivative
+Derive
+Whole
+Positive
+Negative
+Tangent
+Radius
+Diameter
+Surface
+Nonagon
+Nonagonal
+Decagon
+Decagonal
+Hendecagon
+Dodecagon
+Polygon
+Position
+Point
+Coordinate
+Inverse
+Reverse
+Rank
+Sort
+Real
+Formula
+Calculator
+Diagram
+Set
+Chart
+Graph
+Transform
+Transformation
+Inflection
+Inflect
+Reflection
+Reflect
+Sign
+Magnitude
+Modulus
+Absolute
+Constant
+Equal
+Equivalent
+Proposition
+Data
+Collection
+Arc
+Log
+Logarithms
+Analysis
+Vector
+Vectors
+Velocity
+Acceleration
+Length
+Width
+Scalar
+Variable
+Parameter
+Cone
+Cones
+Cube
+Cubes
+Cuboid
+Cuboids
+Cylinder
+Cylinders
+Tetrahedron
+Tetrahedrons
+Pyramid
+Pyramids
+Sphere
+Spheres
+Prism
+Prisms
+Pi
+Matrix
+Matrices
+Identity
+Identities
+Expression
+Entity
+Entities
+Cosine
+Sine
+Pair
+Pairs
+Single
+Singular
+Double
+Triple
+Quadruple
+Quintuple
+Solution
+Gradient
+Integral
+Integrals
+Plane
+Dimension
+Dimensions
+Upper
+Lower
+Bound
+Bounds
+Boundary
+Boundaries
+Contradiction
+Proof
+Rational
+Irrational
+Standard
+Deviation
+Variance
+Proposition
+Validity
+Discrete
+Notation
+Quantity
+Quantities
+Quality
+Domain
+Relationship
+Relation
+Relations
+Function
+Functions
+Total
+Right
+Left
+Symmetric
+Asymmetric
+Asymptote
+Asymptotic
+Complex
+Complexity
+Increase
+Decrease
+Intercept
+Intersect
+Intersects
+Intersection
+Row
+Rows
+Column
+Columns
+Concave
+Convex
+Sigma
+Logic
+Rule
+Mental
+Math
+Maths
+Mathematics
+Mathematical
+Mathematician
+Mathematicians
+Bracket
+Brackets
+Distance
